Created Home and variables file to separate from App
