+++ src/App.css
@@ -69,25 +69,6 @@ h3 {
   color: var(--purple1);
 }
 
-/* calendar month */
-h4 {
-  font-family: var(--regularFont);
-  font-weight: 300;
-  font-size: 40pt;
-  color: var(--purple1);
-  text-transform: uppercase;
-}
-
-/* calendar days of the week */
-h5 {
-  font-family: var(--iconFont);
-  transform: scale(1, 0.9);
-  font-weight: 600;
-  font-size: 17pt;
-  color: var(--darkPurple2);
-  text-transform: uppercase;
-}
-
 img {
   color: var(--darkPurple2);
   font-size: 4pt;
@@ -114,225 +95,7 @@ img {
   gap: 16px;
 }
 
-.body {
-  background-color: var(--white);
-  /* placeholder */
-
-  display: flex;
-  flex-direction: row;
-
-  height: 100vh;
-  height: 100vh;
-}
-
-.leftFrame {
-  background-color: var(--lightGray);
-
-  display: flex;
-  flex-direction: column;
-  gap: 16px;
-
-  /* minus padding */
-  padding-left: 32px;
-  padding-right: 32px;
-  padding-top: 32px;
-  padding-bottom: 32px;
-}
-
-.rightFrame {
-  background-color: var(--white);
-
-  width: 100%;
-
-  display: flex;
-  flex-direction: column;
-
-  padding-left: 38px;
-  padding-right: 38px;
-  padding-top: 32px;
-  padding-bottom: 32px;
-}
-
-.calendarHeader {
-  background-color: var(--white);
-
-  display: flex;
-  flex-direction: column;
-}
-
-.calendarNav {
-  background-color: var(--white);
-
-  display: flex;
-  flex-direction: row;
-
-  align-items: center;
-  justify-content: space-between;
-  justify-content: space-between;
-
-  height: 72px;
-  margin-left: 30px;
-  margin-right: -30px;
-  margin-left: 40px;
-  margin-right: -20px;
-}
-
-.weekLabel {
-  background-color: var(--white);
-
-  display: flex;
-  flex-direction: column;
-  flex-grow: 70;
-  flex-grow: 70;
-
-  justify-content: space-evenly;
-
-  height: 72px;
-}
-
-.daysOfWeek {
-  background-color: var(--white);
-  display: flex;
-  flex-direction: row;
-
-  justify-content: space-around;
-  gap: 5px;
-}
-
-.dateStyle {
-  font-size: 14pt;
-  font-weight: 400;
-  color: var(--darkPurple1);
-}
-
-.timeStyle {
-  font-size: 12pt;
-  font-weight: 400;
-  word-spacing: -5px;
-  color: var(--darkPurple1);
-}
-
-.calendarBodyContainer {
-  width: 100%;
-  height: 100%;
-  display: flex;
-  flex-direction: row;
-
-  align-items: flex-start;
-}
-
-.timeMarkings {
-  background-color: var(--white);
-
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  padding-right: 10px;
-
-  width: 54px;
-  /* minus padding */
-  height: 95.5%;
-
-  text-align: end;
-}
-
-.eventContainer {
-  background-color: var(--purpleEventBg);
-
-  z-index: 2;
-  border-radius: 4px;
-
-  grid-column: monStart;
-  grid-row: AM10start / PM1start;
-
-  display: flex;
-  flex-direction: row;
-
-  align-items: flex-start;
-  overflow: hidden;
-  overflow: hidden;
-
-  gap: 6px;
-}
-
-.rect {
-  background-color: var(--purpleEvent);
-  border-radius: 4px 0 0 4px;
-  width: 5px;
-  height: 100%;
-}
-
-.eventInfo {
-  width: 100%;
-  width: 100%;
-  height: auto;
-
-  display: flex;
-  flex-direction: column;
-  flex-wrap: nowrap;
-
-  padding-right: 5px;
-  padding-top: 5px;
-  padding-bottom: 5px;
-
-  font-family: var(--regularFont);
-  font-weight: 300;
-  font-size: 12pt;
-  color: var(--white);
-  transform: scale(1, 0.95);
-}
-
-.eventTopBar {
-  display: flex;
-  flex-direction: row;
-
-  justify-content: space-between;
-}
-
-.eventTime {
-  display: flex;
-  flex-direction: row;
-
-  font-size: 11pt;
-
-  gap: 3px;
-}
-
-.eventName {
-  font-size: 14pt;
-  font-weight: 500;
-  text-transform: uppercase;
-}
-
-.calendarGrid {
-  background-color: var(--lightGray);
-
-  display: grid;
-  width: 100%;
-  height: 100%;
-
-  grid-template-rows: [AM7start] 1fr [AM730start] 1fr [AM8start]1fr [AM830start] 1fr [AM9start]1fr [AM930start]1fr [AM10start]1fr [AM1030start]1fr [AM11start]1fr [AM1130start]1fr [PM12start]1fr[PM1230start]1fr [PM1start]1fr [PM130start]1fr [PM2start]1fr [PM230start]1fr [PM3start]1fr [PM330start]1fr [PM4start]1fr [PM430start]1fr [PM5start]1fr [PM530start]1fr [PM6start]1fr [PM630start]1fr [PM7start]1fr [PM730start]1fr [PM8start]1fr [PM830start]1fr [PM9start]1fr [PM930start] 1fr [PM10start];
-  grid-template-columns: [sunStart]3fr [monStart]3fr [tueStart]3fr [wedStart]3fr [thuStart]3fr [friStart]3fr [satStart]3fr [weekEnd];
-
-  overflow: hidden;
-  border: 2px solid var(--lightGray);
-}
-
-.gridLine {
-  background-color: var(--white);
-
-  z-index: 1;
-  margin-right: 1px;
-  margin-left: 1px;
-  margin-top: 1px;
-  margin-bottom: 1px;
-  margin-right: 1px;
-  margin-left: 1px;
-  margin-top: 1px;
-  margin-bottom: 1px;
-}
-
-.subHeader {
+.leftHead {
   display: flex;
   flex-direction: row;
 
@@ -423,23 +186,6 @@ img {
   flex-direction: row;
   align-items: center;
   gap: 16px;
-}
-.xButtonContainer {
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  width: 250px; 
-  height: 0px; 
-  justify-content: flex-end;
-}
-.sendSaveContainer {
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  width: 250px;
-  height: 0px;
-  gap: 5px;
-  justify-content: flex-end;
 }
 .dateTimeContainer {
   display: flex;
@@ -547,26 +293,36 @@ button {
   margin-right: 0;
 }
 
-.calendarVariation {
-  width: 20px;
-  height: 20px;
-  font-size: 8pt;
-  margin-right: 0;
-}
-
 .filterButton {
   background-color: var(--purple1);
   width: 25px;
   height: 25px;
 }
 
-.crossButton {
+.xButtonContainer {
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  width: 250px; 
+  height: 0px; 
+  justify-content: flex-end;
+}
+.xButton {
   background-color: var(--white);
   /* placeholder */
   width: 10px;
   height: 10px;
 }
 
+.sendSaveContainer {
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  width: 250px;
+  height: 0px;
+  gap: 5px;
+  justify-content: flex-end;
+}
 .sendSaveButtons {
   background-color: var(--white);
   width: 15px;
