+++ src/variables.js
@@ -1,0 +1,4 @@
+export const getCSSVar = (variable) => {
+    return getComputedStyle(document.documentElement).getPropertyValue(variable).trim();
+  }
+
